api/requirements.txt
flask
flask-cors
openai
cerebras-cloud-sdk
python-dotenv
google-generativeai